refactor: extract constants, debug globals, and toolbar messages from ruby-tab
Extract shared constants (FONT_SIZES, DEFAULT_FONT_SIZE, localStorage keys)
into ruby-tab/constants.js, debug globals helper into ruby-tab/debug-globals.js,
and toolbar i18n messages into ruby-toolbar/messages.js.

Part of #275 (Phase 1)
