+++ packages/scratch-gui/src/components/ruby-toolbar/messages.js
@@ -1,0 +1,98 @@
+// === Smalruby: This file is Smalruby-specific (Ruby toolbar i18n messages) ===
+
+import {defineMessages} from 'react-intl';
+
+const messages = defineMessages({
+    executeLine: {
+        id: 'gui.rubyToolbar.executeLine',
+        defaultMessage: 'Execute current line',
+        description: 'Tooltip for execute line button'
+    },
+    stopExecution: {
+        id: 'gui.rubyToolbar.stopExecution',
+        defaultMessage: 'Stop execution',
+        description: 'Tooltip for stop execution button'
+    },
+    search: {
+        id: 'gui.rubyToolbar.search',
+        defaultMessage: 'Search',
+        description: 'Tooltip for search button'
+    },
+    undo: {
+        id: 'gui.rubyToolbar.undo',
+        defaultMessage: 'Undo',
+        description: 'Tooltip for undo button'
+    },
+    redo: {
+        id: 'gui.rubyToolbar.redo',
+        defaultMessage: 'Redo',
+        description: 'Tooltip for redo button'
+    },
+    prevSprite: {
+        id: 'gui.rubyToolbar.prevSprite',
+        defaultMessage: 'Previous sprite',
+        description: 'Tooltip for previous sprite button'
+    },
+    nextSprite: {
+        id: 'gui.rubyToolbar.nextSprite',
+        defaultMessage: 'Next sprite',
+        description: 'Tooltip for next sprite button'
+    },
+    commandPlaceholder: {
+        id: 'gui.rubyToolbar.commandPlaceholder',
+        defaultMessage: 'Search sprites by name',
+        description: 'Placeholder for command input'
+    },
+    download: {
+        id: 'gui.rubyToolbar.download',
+        defaultMessage: 'Download Ruby code',
+        description: 'Tooltip for download button'
+    },
+    aiAssistant: {
+        id: 'gui.rubyToolbar.aiAssistant',
+        defaultMessage: 'Smalruby Teacher (Gemini)',
+        description: 'Tooltip for AI assistant button'
+    },
+    furiganaOn: {
+        id: 'gui.rubyToolbar.furiganaOn',
+        defaultMessage: 'Hide furigana',
+        description: 'Tooltip for furigana toggle button when ON'
+    },
+    furiganaOff: {
+        id: 'gui.rubyToolbar.furiganaOff',
+        defaultMessage: 'Show furigana',
+        description: 'Tooltip for furigana toggle button when OFF'
+    },
+    autoCorrectOn: {
+        id: 'gui.rubyToolbar.autoCorrectOn',
+        defaultMessage: 'Disable auto-correct',
+        description: 'Tooltip for auto-correct toggle button when ON'
+    },
+    autoCorrectOff: {
+        id: 'gui.rubyToolbar.autoCorrectOff',
+        defaultMessage: 'Enable auto-correct',
+        description: 'Tooltip for auto-correct toggle button when OFF'
+    },
+    moreOptions: {
+        id: 'gui.rubyToolbar.moreOptions',
+        defaultMessage: 'More options',
+        description: 'Tooltip for three-dot menu button'
+    },
+    autoCorrectSettings: {
+        id: 'gui.rubyToolbar.autoCorrectSettings',
+        defaultMessage: 'Auto-Correct Settings',
+        description: 'Label for auto-correct settings menu item'
+    },
+    saveRubyScript: {
+        id: 'gui.rubyToolbar.saveRubyScript',
+        defaultMessage: 'Save Ruby script',
+        description: 'Label for save Ruby script menu item'
+    },
+    stage: {
+        id: 'gui.rubyToolbar.stage',
+        defaultMessage: 'Stage',
+        description: 'Name for stage target'
+    }
+});
+
+export default messages;
