+++ packages/scratch-gui/src/components/ruby-toolbar/ruby-toolbar.jsx
@@ -1,9 +1,10 @@
 import React, {useState, useCallback, useRef, useEffect} from 'react';
 import PropTypes from 'prop-types';
-import {defineMessages, useIntl} from 'react-intl';
+import {useIntl} from 'react-intl';
 import VM from '@smalruby/scratch-vm';
 
 import styles from './ruby-toolbar.css';
+import messages from './messages.js';
 
 import iconPlay from './icon--play.svg';
 import iconStop from './icon--stop.svg';
@@ -16,99 +17,6 @@ import iconDownload from './icon--download.svg';
 import iconAI from './icon--ai.svg';
 import iconFurigana from './icon--furigana.svg';
 import iconAutoCorrect from './icon--auto-correct.svg';
-
-const messages = defineMessages({
-    executeLine: {
-        id: 'gui.rubyToolbar.executeLine',
-        defaultMessage: 'Execute current line',
-        description: 'Tooltip for execute line button'
-    },
-    stopExecution: {
-        id: 'gui.rubyToolbar.stopExecution',
-        defaultMessage: 'Stop execution',
-        description: 'Tooltip for stop execution button'
-    },
-    search: {
-        id: 'gui.rubyToolbar.search',
-        defaultMessage: 'Search',
-        description: 'Tooltip for search button'
-    },
-    undo: {
-        id: 'gui.rubyToolbar.undo',
-        defaultMessage: 'Undo',
-        description: 'Tooltip for undo button'
-    },
-    redo: {
-        id: 'gui.rubyToolbar.redo',
-        defaultMessage: 'Redo',
-        description: 'Tooltip for redo button'
-    },
-    prevSprite: {
-        id: 'gui.rubyToolbar.prevSprite',
-        defaultMessage: 'Previous sprite',
-        description: 'Tooltip for previous sprite button'
-    },
-    nextSprite: {
-        id: 'gui.rubyToolbar.nextSprite',
-        defaultMessage: 'Next sprite',
-        description: 'Tooltip for next sprite button'
-    },
-    commandPlaceholder: {
-        id: 'gui.rubyToolbar.commandPlaceholder',
-        defaultMessage: 'Search sprites by name',
-        description: 'Placeholder for command input'
-    },
-    download: {
-        id: 'gui.rubyToolbar.download',
-        defaultMessage: 'Download Ruby code',
-        description: 'Tooltip for download button'
-    },
-    aiAssistant: {
-        id: 'gui.rubyToolbar.aiAssistant',
-        defaultMessage: 'Smalruby Teacher (Gemini)',
-        description: 'Tooltip for AI assistant button'
-    },
-    furiganaOn: {
-        id: 'gui.rubyToolbar.furiganaOn',
-        defaultMessage: 'Hide furigana',
-        description: 'Tooltip for furigana toggle button when ON'
-    },
-    furiganaOff: {
-        id: 'gui.rubyToolbar.furiganaOff',
-        defaultMessage: 'Show furigana',
-        description: 'Tooltip for furigana toggle button when OFF'
-    },
-    autoCorrectOn: {
-        id: 'gui.rubyToolbar.autoCorrectOn',
-        defaultMessage: 'Disable auto-correct',
-        description: 'Tooltip for auto-correct toggle button when ON'
-    },
-    autoCorrectOff: {
-        id: 'gui.rubyToolbar.autoCorrectOff',
-        defaultMessage: 'Enable auto-correct',
-        description: 'Tooltip for auto-correct toggle button when OFF'
-    },
-    moreOptions: {
-        id: 'gui.rubyToolbar.moreOptions',
-        defaultMessage: 'More options',
-        description: 'Tooltip for three-dot menu button'
-    },
-    autoCorrectSettings: {
-        id: 'gui.rubyToolbar.autoCorrectSettings',
-        defaultMessage: 'Auto-Correct Settings',
-        description: 'Label for auto-correct settings menu item'
-    },
-    saveRubyScript: {
-        id: 'gui.rubyToolbar.saveRubyScript',
-        defaultMessage: 'Save Ruby script',
-        description: 'Label for save Ruby script menu item'
-    },
-    stage: {
-        id: 'gui.rubyToolbar.stage',
-        defaultMessage: 'Stage',
-        description: 'Name for stage target'
-    }
-});
 
 const RubyToolbar = props => {
     const intl = useIntl();
