+++ packages/scratch-gui/src/containers/ruby-tab/constants.js
@@ -1,0 +1,15 @@
+// === Smalruby: This file is Smalruby-specific (Ruby tab constants) ===
+
+const FONT_SIZES = [8, 10, 12, 14, 16, 18, 20, 24, 28, 32, 36, 40, 48];
+const DEFAULT_FONT_SIZE = 16;
+const FURIGANA_ENABLED_KEY = 'smalruby:furiganaEnabled';
+const AUTO_CORRECT_ENABLED_KEY = 'smalruby:autoCorrectEnabled';
+const AUTO_CORRECT_SETTINGS_KEY = 'smalruby:autoCorrectSettings';
+
+export {
+    FONT_SIZES,
+    DEFAULT_FONT_SIZE,
+    FURIGANA_ENABLED_KEY,
+    AUTO_CORRECT_ENABLED_KEY,
+    AUTO_CORRECT_SETTINGS_KEY
+};
